+++ src/main/kotlin/com/mojh/cms/security/dto/request/LoginRequest.kt
@@ -1,4 +1,4 @@
-package com.mojh.cms.member.dto
+package com.mojh.cms.security.dto.request
 
 import javax.validation.constraints.NotBlank
 import javax.validation.constraints.Size
@@ -11,14 +11,4 @@ data class LoginRequest(
     @field:NotBlank
     @field:Size(min = 6, max = 20)
     val password: String,
-)
-
-data class LogoutRequest(
-    @field:NotBlank
-    val refreshToken: String
-)
-
-data class ReissueTokenRequest(
-    @field:NotBlank
-    val refreshToken: String
 )
