+++ src/main/kotlin/com/mojh/cms/security/dto/request/LogoutRequest.kt
@@ -1,0 +1,8 @@
+package com.mojh.cms.security.dto.request
+
+import javax.validation.constraints.NotBlank
+
+data class LogoutRequest(
+    @field:NotBlank
+    val refreshToken: String
+)
