+++ src/main/kotlin/com/mojh/cms/security/dto/request/ReissueTokenRequest.kt
@@ -1,0 +1,8 @@
+package com.mojh.cms.security.dto.request
+
+import javax.validation.constraints.NotBlank
+
+data class ReissueTokenRequest(
+    @field:NotBlank
+    val refreshToken: String
+)
