+++ src/main/kotlin/com/mojh/cms/security/service/AuthService.kt
@@ -2,9 +2,9 @@ package com.mojh.cms.security.service
 
 import com.mojh.cms.common.exception.CustomException
 import com.mojh.cms.common.exception.ErrorCode.*
-import com.mojh.cms.member.dto.LoginRequest
 import com.mojh.cms.member.repository.MemberRepository
 import com.mojh.cms.security.dto.TokensResponse
+import com.mojh.cms.security.dto.request.LoginRequest
 import com.mojh.cms.security.jwt.JwtTokenUtils
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.security.crypto.password.PasswordEncoder
